requirements.txt
torch>=1.7.0
torchaudio>=0.7.0
numpy>=1.19.0
pandas>=1.1.0
librosa>=0.8.0
soundfile>=0.10.0
tqdm>=4.50.0
matplotlib>=3.3.0
scipy>=1.5.0
pyyaml>=5.3.0
tensorboard>=2.4.0
transformers>=4.0.0
datasets>=1.0.0
phonemizer>=3.2.1
espeak-ng>=1.51 